[KYUUBI #3648][SUBTASK][FEATURE][UI] Fix Unit Test
